Add type constraints, remove unnecessary casts, etc

+++ src/backend/src/person/controller/person_delete.rs
@@ -2,6 +2,6 @@ use crate::person::{person_manager::PersonManager, person_types::Person};
 use ic_cdk::update;
 
 #[update]
-pub fn person_delete(id: u32) -> Result<Person, String> {
+pub fn person_delete(id: i64) -> Result<Person, String> {
     PersonManager::delete(id)
 }
